Commit 2
Commit 2

+++ src/routes/carts.router.js
@@ -8,6 +8,7 @@ const logger = pino({},pino.multistream(streams))
 const router = Router();
 
 router.get('/:cid/products',async(req,res)=>{
+   if(!req.session.user) return res.redirect('/login')
    logger.info(`Coneccion recibida en ' /api/carts/:cid/products ' con metodo GET`)
    let cid = req.params.cid
    if(isNaN(cid)) return res.status(400).send({error:"El valor no es numerico"})
@@ -17,14 +18,15 @@ router.get('/:cid/products',async(req,res)=>{
 })
 
 router.post('/',async(req,res)=>{
+   if(!req.session.user) return res.redirect('/login')
    logger.info(`Coneccion recibida en ' /api/carts/ ' con metodo POST`)
    let create = await services.cartsService.createCart()
    res.send(`El id de su carrito es ${create}`)
 })
 router.post('/products/:pid',async(req,res)=>{
-   logger.info(`Coneccion recibida en ' /api/carts/:cid/products/:pid ' con metodo POST`)
+   if(!req.session.user) return res.redirect('/login')
+   logger.info(`Coneccion recibida en ' /api/cartsproducts/:pid ' con metodo POST`)
    let info = {pid:req.params.pid,cid:req.session.user.cartID}
-   console.log(info)
    if(isNaN(info.pid)) return res.status(400).send({error:"El valor no es numerico o no existe"})
    await services.cartsService.addProduct(info)
    let list = await services.cartsService.getCartProducts(req.session.user.cartID)
@@ -34,6 +36,7 @@ router.post('/products/:pid',async(req,res)=>{
 
 
 router.delete('/:cid',async(req,res)=>{
+   if(!req.session.user) return res.redirect('/login')
    logger.info(`Coneccion recibida en ' /api/carts/:cid ' con metodo DELETE`)
    let cid = req.params.cid
    if(isNaN(cid)) return res.status(400).send({error:"El valor no es numerico"})
@@ -42,7 +45,8 @@ router.delete('/:cid',async(req,res)=>{
 })
 
 router.post('/delete/:pid',async(req,res)=>{
-   logger.info(`Coneccion recibida en ' /api/carts/products/:pid ' con metodo DELETE`)
+   if(!req.session.user) return res.redirect('/login')
+   logger.info(`Coneccion recibida en ' /api/carts/delete/:pid ' con metodo DELETE`)
    let cid = {pid:req.params.pid,cid:req.session.user.cartID}
    if(isNaN(cid.cid)) return res.status(400).send({error:"El valor no es numerico"})
    let deleten = await services.cartsService.deleteByCidAndPid(cid)
@@ -50,9 +54,15 @@ router.post('/delete/:pid',async(req,res)=>{
 })
 
 router.post('/endshop',async(req,res)=>{
-   let user = req.session.user
-   let result = await services.cartsService.endShop(req.session.user,req.session.user.cartID)
-   res.render('endShop',{user})
+   if(!req.session.user) return res.redirect('/login')
+   logger.info(`Coneccion recibida en ' /api/carts/endshop ' con metodo post`)
+   try {
+      let user = req.session.user
+      let result = await services.cartsService.endShop(req.session.user,req.session.user.cartID)
+      res.render('endShop',{user})
+   } catch (error) {
+      logger.error(`Hay un error ${error}`)
+   }
 })
